requirements.txt
streamlit>=1.28.0
google-generativeai>=0.8.0
python-docx>=1.1.0
python-dotenv>=1.0.0
